fix(button): Move utils into a separate file to fix use before define
