+++ modules/button/react/lib/IconButton.tsx
@@ -1,9 +1,9 @@
 import * as React from 'react';
-import {getButtonStyle} from './ButtonBase';
 import styled from '@emotion/styled';
 import isPropValid from '@emotion/is-prop-valid';
 import {IconButtonVariant, IconButtonSize} from './types';
 import {iconButtonStyles} from './ButtonStyles';
+import {getButtonStyle} from './utils';
 import {colors} from '@workday/canvas-kit-react-core';
 import {SystemIcon} from '@workday/canvas-kit-react-icon';
 import {focusRing} from '@workday/canvas-kit-react-common';
@@ -39,6 +39,30 @@ export interface IconButtonProps extends React.ButtonHTMLAttributes<HTMLButtonEl
    * Callback that fires when a button changes toggled states
    */
   onToggleChange?: (toggled: boolean | undefined) => void;
+}
+
+function getFillSelector(fillColor: string): CSSObject {
+  return {
+    '&:focus span .wd-icon-fill, &:hover span .wd-icon-fill, span .wd-icon-fill': {
+      fill: fillColor,
+    },
+  };
+}
+
+function getBackgroundSelector(fillColor: string): CSSObject {
+  return {
+    '&:hover span .wd-icon-background, span .wd-icon-background': {
+      fill: fillColor,
+    },
+  };
+}
+
+function getAccentSelector(fillColor: string): CSSObject {
+  return {
+    '&:focus span .wd-icon-accent, &:hover span .wd-icon-accent, span .wd-icon-accent': {
+      fill: fillColor,
+    },
+  };
 }
 
 export const IconButtonCon = styled('button', {
@@ -148,30 +172,6 @@ export const IconButtonCon = styled('button', {
     }
   }
 );
-
-function getFillSelector(fillColor: string): CSSObject {
-  return {
-    '&:focus span .wd-icon-fill, &:hover span .wd-icon-fill, span .wd-icon-fill': {
-      fill: fillColor,
-    },
-  };
-}
-
-function getBackgroundSelector(fillColor: string): CSSObject {
-  return {
-    '&:hover span .wd-icon-background, span .wd-icon-background': {
-      fill: fillColor,
-    },
-  };
-}
-
-function getAccentSelector(fillColor: string): CSSObject {
-  return {
-    '&:focus span .wd-icon-accent, &:hover span .wd-icon-accent, span .wd-icon-accent': {
-      fill: fillColor,
-    },
-  };
-}
 
 export default class IconButton extends React.Component<IconButtonProps> {
   public static Variant = IconButtonVariant;
