+++ modules/button/react/lib/TextButton.tsx
@@ -1,5 +1,6 @@
 import * as React from 'react';
-import {ButtonBaseLabel, ButtonLabelIcon, getButtonStyle} from './ButtonBase';
+import {ButtonBaseLabel, ButtonLabelIcon} from './ButtonBase';
+import {getButtonStyle} from './utils';
 import styled from '@emotion/styled';
 import isPropValid from '@emotion/is-prop-valid';
 import {ButtonSize, IconPosition, TextButtonVariant} from './types';
